+++ straight_skeleton/src/main/kotlin/adder.kt
@@ -1,0 +1,11 @@
+package kotlindemo
+
+class Adder
+
+fun add(x: Int, y: Int): Int {
+    return x+y
+}
+
+fun split(str: String): String {
+    return str
+}