+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Added a batch runner
+* Added a batch runner & a set of test cases
 * Implemented plaintext/xz-compressed coordinates reading & loading
 * Added autonomous single file runner
 * Added some simple polygons which yield incorrect behaviour